Move integration tests to separate class
