+++ pensieve/tests/test_analysis.py
@@ -1,22 +1,12 @@
 import datetime as dt
 from datetime import timedelta
 import json
-import google
-import mock
 import os
 import pytz
 import pytest
 
-from mozanalysis.metrics import Metric, DataSource, agg_sum
-
-from pensieve.analysis import Analysis, BigQueryClient
-from pensieve.experimenter import Experiment, Variant
-
-test_project = "pensieve-integration-test"
-test_dataset = "test"
-static_dataset = (
-    "test_data"  # contains the tables filled with test data required to run metrics analysis
-)
+from pensieve.analysis import Analysis
+from pensieve.experimenter import Experiment
 
 
 @pytest.fixture
@@ -148,79 +138,3 @@ def test_regression_20200316():
     experiment = Experiment.from_dict(json.loads(experiment_json))
     analysis = Analysis("test", "test", experiment)
     analysis.run(current_date=dt.datetime(2020, 3, 16), dry_run=True)
-
-
-def bq_execute(cls, *args, **kwargs):
-    # to use the test project and dataset, we need to change the SQL query generated by mozanalysis
-    query = args[0].replace("moz-fx-data-shared-prod", test_project)
-    query = query.replace("telemetry", "test_data")
-
-    dataset = google.cloud.bigquery.dataset.DatasetReference.from_string(
-        cls.dataset, default_project=cls.project,
-    )
-
-    destination_table = None
-
-    if len(args) > 1:
-        destination_table = args[1]
-
-    kwargs = {}
-    if destination_table:
-        kwargs["destination"] = dataset.table(destination_table)
-        kwargs["write_disposition"] = google.cloud.bigquery.job.WriteDisposition.WRITE_TRUNCATE
-    config = google.cloud.bigquery.job.QueryJobConfig(default_dataset=dataset, **kwargs)
-    job = cls.client.query(query, config)
-    # block on result
-    return job.result(max_results=1)
-
-@pytest.mark.integration
-def test_metrics():
-    experiment = Experiment(
-        slug="test-experiment",
-        type="rollout",
-        start_date=dt.datetime(2020, 3, 30, tzinfo=pytz.utc),
-        end_date=dt.datetime(2020, 6, 1, tzinfo=pytz.utc),
-        proposed_enrollment=7,
-        variants=[
-            Variant(is_control=False, slug="branch1", ratio=0.5),
-            Variant(is_control=True, slug="branch2", ratio=0.5),
-        ],
-        normandy_slug="test-experiment",
-    )
-
-    # todo create class and move there
-    project_id = os.getenv("GOOGLE_PROJECT_ID")
-
-    if project_id is None:
-        print("GOOGLE_PROJECT_ID is not set.")
-        assert False
-
-    with mock.patch.object(BigQueryClient, 'execute', new=bq_execute):
-        analysis = Analysis(project_id, test_dataset, experiment)
-
-        test_clients_daily = DataSource(
-            name="clients_daily", from_expr=f"`{project_id}.test_data.clients_daily`",
-        )
-
-        test_active_hours = Metric(
-            name="active_hours", data_source=test_clients_daily, select_expr=agg_sum("active_hours_sum")
-        )
-
-        analysis.STANDARD_METRICS = [test_active_hours]
-
-        analysis.run(current_date=dt.datetime(2020, 4, 12), dry_run=False)
-
-    client = google.cloud.bigquery.client.Client(project_id)
-    query_job = client.query(f"""
-        SELECT
-          COUNT(*) as count
-        FROM `{project_id}.{test_dataset}.test_experiment_week_1`
-    """)
-
-    result = query_job.result()
-
-    for row in result:
-        assert row["count"] == 2
-
-    # todo: remove tables that are generated, but leave non-generated
-
